auth-service/server_python/requirements.txt
fastapi
uvicorn
pydantic
motor
bcrypt
pyjwt
python-multipart
